+++ app/src/main/java/com/aad/esei/uvigo/core/Repostaje.java
@@ -2,22 +2,22 @@ package com.aad.esei.uvigo.core;
 
 public class Repostaje extends Gasto{
 
-    private float precio_litro;
+    private double precio_litro;
     private int litros_antes;
     private int litros_repostados;
     private int km_actuales;
 
 
-    public Repostaje(Categoria_Gasto categoria, float precio_litro, int litros_antes, int litros_repostados,int km_actuales, String fecha, String titulo) {
+    public Repostaje(Categoria_Gasto categoria, double precio_litro, int litros_antes, int litros_repostados,int km_actuales, String fecha, String titulo) {
 
-        super(categoria, ((float) precio_litro/litros_antes), fecha, titulo);
+        super(categoria, ((double) precio_litro/litros_antes), fecha, titulo);
         this.precio_litro = precio_litro;
         this.litros_antes = litros_antes;
         this.litros_repostados = litros_repostados;
         this.km_actuales = km_actuales;
     }
 
-    public float getPrecio_litro() {
+    public double getPrecio_litro() {
         return precio_litro;
     }
 
@@ -33,7 +33,7 @@ public class Repostaje extends Gasto{
         return km_actuales;
     }
 
-    public void setPrecio_litro(float precio_litro) {
+    public void setPrecio_litro(double precio_litro) {
         this.precio_litro = precio_litro;
     }
